requirements.txt
fpdf2
openai
python-dotenv
python-fasthtml
